backend/requirements.txt
fastapi==0.110.0
langchain==0.1.10
langchain_community==0.0.25
langchain_google_genai==0.0.9
numpy==1.26.4
python-dotenv==1.0.1
scikit_learn==1.2.2
transformers==4.39.0
